+++ src/RedisPubSubAdapter.js
@@ -73,6 +73,24 @@ class RedisPubSubAdapter {
   publish(channel, message) {
     this.client.publish(channel, Utils.serializeMessagePayload(message));
   }
+
+  /**
+   * Publish multiple messages to a channel.
+   *
+   * @param {string} channel
+   * @param {*[]} messages
+   * @example
+   * let messages = [
+   *   'message 1',
+   *   'message 2',
+   * ];
+   * adapter.publishBatch('my_channel', messages);
+   */
+  publishBatch(channel, messages) {
+    for (let message of messages) {
+      this.publish(channel, message);
+    }
+  }
 }
 
 module.exports = RedisPubSubAdapter;